feat: extend the forked webdav package
