+++ awsfs/remove.go
@@ -2,11 +2,14 @@ package awsfs
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"strings"
 )
 
 func (s Server) RemoveAll(ctx context.Context, path string) error {
+	fmt.Println("RemoveAll: ", path)
+
 	if path = slashClean(path); path == "/" {
 		return os.ErrInvalid
 	}
